added raycasting

+++ neuroevol/Wall.pde
@@ -1,6 +1,6 @@
 class Wall{
   PVector a,b; 
-  Wall(int x1, int y1, int x2, int y2){
+  Wall(float x1, float y1, float x2, float y2){
      this.a = new PVector(x1, y1);
      this.b = new PVector(x2, y2);
    }
